requirements.txt

Pillow
pytesseract
pypdf
python-dotenv 